+++ app/src/main/java/cat/tecnocampus/mobileapps/practica2/HomarMasachsFrancescMeninoSuredaPau/DatabaseController.java
@@ -38,7 +38,7 @@ public class DatabaseController {
     }
 
 
-    public void setGame(String nickname, int points) {
+    public void setGame(String nickname, double points) {
         Game game = new Game();
         game.setNickname(nickname);
         game.setPoints(points);
